ADGM_Corporate_Agent/requirements.txt
gradio>=4.0.0
python-docx>=1.1.0
openai>=1.0.0
anthropic>=0.25.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
nltk>=3.8.1
scikit-learn>=1.3.0
tiktoken>=0.5.0